+++ src/rules/tailwind-no-unnecessary-whitespace.ts
@@ -1,11 +1,15 @@
-import { DEFAULT_ATTRIBUTE_NAMES, DEFAULT_CALLEE_NAMES, DEFAULT_VARIABLE_NAMES } from "src/config/default-config.js";
-
+import {
+  DEFAULT_ATTRIBUTE_NAMES,
+  DEFAULT_CALLEE_NAMES,
+  DEFAULT_VARIABLE_NAMES
+} from "readable-tailwind:config:default-config.js";
 import { getCalleeSchema, getClassAttributeSchema, getVariableSchema } from "readable-tailwind:config:descriptions.js";
 import { getLiteralsByESCallExpression, getLiteralsByESVariableDeclarator } from "readable-tailwind:parsers:es.js";
 import { getAttributesByHTMLTag, getLiteralsByHTMLClassAttribute } from "readable-tailwind:parsers:html.js";
 import { getAttributesByJSXElement, getLiteralsByJSXClassAttribute } from "readable-tailwind:parsers:jsx.js";
 import { getAttributesBySvelteTag, getLiteralsBySvelteClassAttribute } from "readable-tailwind:parsers:svelte.js";
 import { getAttributesByVueStartTag, getLiteralsByVueClassAttribute } from "readable-tailwind:parsers:vue.js";
+import { escapeNestedQuotes } from "readable-tailwind:utils:quotes.js";
 import { splitClasses, splitWhitespaces } from "readable-tailwind:utils:utils.js";
 
 import type { TagNode } from "es-html-parser";
@@ -97,7 +101,7 @@ export const tailwindNoUnnecessaryWhitespace: ESLintRule<Options> = {
           const htmlAttributes = getAttributesByHTMLTag(ctx, htmlTagNode);
 
           for(const htmlAttribute of htmlAttributes){
-            const literals = getLiteralsByHTMLClassAttribute(ctx, htmlAttribute);
+            const literals = getLiteralsByHTMLClassAttribute(ctx, htmlAttribute, classAttributes);
             lintLiterals(ctx, literals);
           }
         }
@@ -106,9 +110,15 @@ export const tailwindNoUnnecessaryWhitespace: ESLintRule<Options> = {
       // Vue
       if(typeof ctx.sourceCode.parserServices?.defineTemplateBodyVisitor === "function"){
         return {
+          // script tag
           ...callExpression,
           ...variableDeclarators,
-          ...ctx.sourceCode.parserServices.defineTemplateBodyVisitor(vue)
+
+          // bound classes
+          ...ctx.sourceCode.parserServices.defineTemplateBodyVisitor({
+            ...callExpression,
+            ...vue
+          })
         };
       }
 
@@ -158,10 +168,15 @@ function lintLiterals(ctx: Rule.RuleContext, literals: Literal[]) {
 
     const classes = splitClassesKeepWhitespace(literal, allowMultiline);
 
+    const escapedClasses = escapeNestedQuotes(
+      classes.join(""),
+      literal.openingQuote ?? "\""
+    );
+
     const fixedClasses = [
       literal.openingQuote ?? "",
       literal.type === "TemplateLiteral" && literal.closingBraces ? literal.closingBraces : "",
-      ...classes,
+      escapedClasses,
       literal.type === "TemplateLiteral" && literal.openingBraces ? literal.openingBraces : "",
       literal.closingQuote ?? ""
     ].join("");