+++ src/utils/quotes.ts
@@ -1,0 +1,6 @@
+import type { LiteralValueQuotes } from "readable-tailwind:types:ast.js";
+
+
+export function escapeNestedQuotes(content: string, surroundingQuotes: LiteralValueQuotes): string {
+  return content.replace(new RegExp(surroundingQuotes, "g"), `\\${surroundingQuotes}`);
+}